% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beamys_hindei_range}
\alias{Beamys_hindei_range}
\title{Expert range data for Beamys_hindei}
\format{A SpatialPolygonDataFrame object with 2 versions of the expert range map}
\usage{
Beamys_hindei_range
}
\description{
Expert range data extracted from the Map Of Life (\url{mol.org}).
}
\author{
Adam M. Wilson
}
\references{
\url{MOL.org}
}
\keyword{datasets}

