% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{Make maxent's cumulative output from raw output and (optionally) apply a threshold}
\usage{
cumulative(x, threshold = NULL)
}
\arguments{
\item{x}{raster* object of continuous (raw) maxent output}

\item{threshold}{\code{scalar} value to threshold the continuous map into binary predictions}
}
\description{
The function first converts the raw maxent output to cumulative (by sorting and calculating the cumulative sum).
Then the (optional) threshold is applied to return a binary map of the estimated range.
If \code{threshold} is \code{NULL}, the cumulative output is returned.
}

