% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbbi.R
\name{fbbi}
\alias{fbbi}
\title{Factor Bias Back In (FBBI)}
\usage{
fbbi(prior, maxent, ...)
}
\arguments{
\item{prior}{raster* object of spatial priors}

\item{maxent}{raster* object of maxent output}

\item{...}{additional functions to be passed to \code{\link[raster]{writeRaster}}}
}
\description{
Calculate minxent layer using maxent output  (from run with bias layer) and prior (i.e., the bias layer)
}

