% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leucadendron_lanigerum_range}
\alias{Leucadendron_lanigerum_range}
\title{Expert range data for Leucadendron lanigerum}
\format{A Raster layer object with 1s for the expert range, 0s for the rest of the domain, and NAs outside the domain.}
\usage{
Leucadendron_lanigerum_range
}
\description{
Expert range data for Leucadendron lanigerum, a South African Shrub.
}
\author{
Adam M. Wilson
}
\references{
\url{MOL.org}
}
\keyword{datasets}

