% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a raster}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{raster* object}

\item{...}{additional functions to be passed to \code{\link[raster]{writeRaster}}}
}
\description{
Divide raster by the sum of all cells.
}

