% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinside.R
\name{pinside}
\alias{pinside}
\title{Calculate the probability inside an expert range given a logistic decay}
\usage{
pinside(dists, parms)
}
\arguments{
\item{dists}{frequency table of unique distance values (output from running \code{\link[raster]{freq}} on the \code{rdist} object).}

\item{parms}{A named vector of parameter values from \code{\link{logistic}} that describe the desired curve.
These include all parameters for \code{\link{logistic}} (\code{rate,skew,shift}).
Optionally, you can add a parameter \code{buffer} which adds a buffer around the range as 'inside.'}
}
\value{
Scalar of the estimated probability inside the range given the specified logistic decay (with optional buffer).
}
\description{
Uses the \code{\link{logistic}} function to estimate the probability inside the range
}

