% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmRaster.R
\name{rmRaster}
\alias{rmRaster}
\title{Really remove raster files}
\usage{
rmRaster(x, verbose = FALSE)
}
\arguments{
\item{x}{a \code{raster*} object to be removed}

\item{verbose}{logical indicating whether to print verbose messages}
}
\description{
The raster package has the sometimes problematic behavior of leaving temporary files after the R object has been removed.  
This function first checks if a temporary file exists on disk and removes it.
}
\author{
Adam M. Wilson

Keith Ma
}
\references{
https://gist.github.com/adammwilson/10180852
}

