% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tinamus_solitarius_env}
\alias{Tinamus_solitarius_env}
\title{Expert range environmental data for the Solitary Tinamou (Tinamus solitarius)}
\format{A rasterStack object}
\usage{
Tinamus_solitarius_env
}
\description{
Environmental data (from WorldClim) for the Solitary Tinamou's range (\url{mol.org}).
Bio1 is Mean Annual Temperature and Bio12 is Mean Annual Precipitation.
}
\author{
Adam M. Wilson
}
\keyword{datasets}

