% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tinamus_solitarius_points}
\alias{Tinamus_solitarius_points}
\title{Point distributional data for the Solitary Tinamou (Tinamus solitarius)}
\format{A SpatialPointsDataFrame object with 171 point locations and the following variables: 
\code{cartodb_id}, \code{type},
\code{provider}, \code{seasonality},
\code{presence}, \code{uncertainty},
and \code{date} collected.}
\usage{
Tinamus_solitarius_points
}
\description{
Point ocurrance data from GBIF and eBird extracted from the Map Of Life (\url{mol.org}).
}
\author{
Adam M. Wilson
}
\references{
\url{MOL.org}
}
\keyword{datasets}

