% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tinamus_solitarius_range}
\alias{Tinamus_solitarius_range}
\title{Expert range data for the Solitary Tinamou (Tinamus solitarius)}
\format{A SpatialPolygonDataFrame object with 2 versions of the expert range map}
\usage{
Tinamus_solitarius_range
}
\description{
Expert range data extracted from the Map Of Life (\url{mol.org}).
}
\author{
Adam M. Wilson
}
\references{
\url{MOL.org}
}
\keyword{datasets}

