% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Graphically display folder tree in terminal}
\usage{
tree(x = NULL)
}
\arguments{
\item{x}{File path to display}
}
\value{
Prints the folder structure
}
\description{
Uses \code{sed} to display formatted folder structure.  Needs sed to be installed on system
 (may not be available on windows)
}
\references{
\url{http://www.cnet.com/news/terminal-fun-options-for-printing-folder-and-subfolder-contents/}
}

